Move ActiveJob::LogSubscriber into its own file
